refactor: remove JS checklist base classes, rewrite as standalone
