+++ amd/src/resource_checklist/resource_checklist_item.js
@@ -16,18 +16,29 @@
 /**
  * Resource checklist item component.
  *
+ * Standalone component handling a single checklist item row.
+ * Manages edit modal, state updates, and DOM updates for one item.
+ *
  * @module mod_bookit/resource_checklist/resource_checklist_item
  * @copyright   2026 ssystems GmbH <oss@ssystems.de>
  * @author      Andreas Rosenthal
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-import BaseChecklistItem from 'mod_bookit/checklist/base_checklist_item';
+import {BaseComponent} from 'core/reactive';
+import Templates from 'core/templates';
+import ModalForm from 'core_form/modalform';
+import {get_string as getString} from 'core/str';
+
+const ITEM_MODAL_FORM = 'mod_bookit\\form\\edit_resource_checklist_item_form';
+const ITEM_TEMPLATE = 'mod_bookit/resource_checklist/resource_checklist_item_row';
+const ITEM_REGION = 'resource-checklist-item-row';
 
 /**
  * Resource checklist item component.
  */
-export default class extends BaseChecklistItem {
+export default class ResourceChecklistItem extends BaseComponent {
+
     /**
      * Component descriptor for debugging.
      *
@@ -38,33 +49,147 @@ export default class extends BaseChecklistItem {
     }
 
     /**
-     * Get configuration object.
+     * Create component.
      *
-     * @return {Object} Configuration
+     * @param {Object} descriptor - Component descriptor
      */
-    getConfig() {
-        return {
-            itemsStateKey: 'checklistitems',
-            itemType: 'checklistitem',
-            itemModalForm: 'mod_bookit\\form\\edit_resource_checklist_item_form',
-            templates: {
-                itemRow: 'mod_bookit/resource_checklist/resource_checklist_item_row',
-            }
-        };
+    create(descriptor) {
+        const itemId = descriptor.element.dataset.itemid || descriptor.element.dataset.bookitItemId;
+        this._editBtnSelector = itemId ? `#edit-item-${itemId}` : null;
+        this.itemData = descriptor.itemData;
     }
 
     /**
-     * Update row with new item data from reactive state.
+     * Get state watchers.
      *
-     * @param {Object} itemData - Updated item data
+     * @return {Array} Empty — item updates are handled via _updateRow() called by the category.
+     */
+    getWatchers() {
+        return [];
+    }
+
+    /**
+     * Attach event listeners once state is ready.
+     */
+    stateReady() {
+        this._attachEventListeners();
+    }
+
+    /**
+     * Attach event listeners.
+     */
+    _attachEventListeners() {
+        if (!this._editBtnSelector) {
+            return;
+        }
+        const editBtn = this.getElement(this._editBtnSelector);
+        if (editBtn) {
+            this.addEventListener(editBtn, 'click', this._handleEdit.bind(this));
+        }
+    }
+
+    /**
+     * Handle edit button click.
+     *
+     * @param {Event} event - Click event
+     */
+    async _handleEdit(event) {
+        event.preventDefault();
+
+        const itemId = event.currentTarget.value || event.currentTarget.dataset.itemid;
+        const modalForm = new ModalForm({
+            formClass: ITEM_MODAL_FORM,
+            args: {id: parseInt(itemId)},
+            modalConfig: {title: await getString('editchecklistitem', 'mod_bookit')},
+            returnFocus: event.currentTarget,
+        });
+
+        modalForm.addEventListener(modalForm.events.FORM_SUBMITTED, (e) => {
+            const response = e.detail;
+
+            // Handle processUpdates format (array of updates from server)
+            if (Array.isArray(response)) {
+                this.reactive.stateManager.processUpdates(response);
+                if (response[0]?.action === 'delete') {
+                    this.reactive.dispatch('itemDeleted', {id: parseInt(response[0].fields.id)});
+                    this.remove();
+                }
+                return;
+            }
+
+            // Handle direct result format
+            if (response.result === 'success') {
+                if (response.action === 'delete') {
+                    this.reactive.dispatch('itemsDeleted', {fields: response.data});
+                    this.remove();
+                } else {
+                    this.reactive.dispatch('itemsUpdated', {fields: response.data});
+                }
+            }
+        });
+
+        modalForm.show();
+    }
+
+    /**
+     * Update the item row DOM with new data.
+     *
+     * @param {Object} itemData - Updated item data from state
      */
     _updateRow(itemData) {
-        super._updateRow(itemData);
+        this.itemData = itemData;
 
-        // Update duedate display cell.
+        if (this.element) {
+            this.element.dataset.itemName = itemData.name || '';
+            this.element.dataset.itemDescription = itemData.description || '';
+            this.element.dataset.itemSortorder = itemData.sortorder || 0;
+        }
+
+        const nameCell = this.element.querySelector('[data-field="item-name"]');
+        if (nameCell) {
+            nameCell.textContent = itemData.name;
+        }
+
+        const descCell = this.element.querySelector('[data-field="item-description"]');
+        if (descCell) {
+            descCell.textContent = itemData.description || '';
+        }
+
+        // Resource-specific: update due date display cell.
         const duedateCell = this.element.querySelector('[data-field="item-duedate"]');
         if (duedateCell) {
             duedateCell.textContent = itemData.duedatedisplay || '-';
         }
     }
+
+    /**
+     * Remove the item element from the DOM.
+     */
+    remove() {
+        if (this.element && this.element.parentNode) {
+            this.element.parentNode.removeChild(this.element);
+        }
+    }
+
+    /**
+     * Render item from template (for dynamically created items).
+     */
+    async _render() {
+        const html = await Templates.render(ITEM_TEMPLATE, this.itemData);
+        const tempDiv = document.createElement('div');
+        tempDiv.innerHTML = html.trim();
+        const itemElement = tempDiv.firstElementChild;
+
+        if (this.element.tagName === 'TBODY') {
+            this.element.appendChild(itemElement);
+        } else {
+            this.element.parentNode.replaceChild(itemElement, this.element);
+        }
+
+        this.element = itemElement;
+        this._editBtnSelector = `#edit-item-${this.itemData.id}`;
+        this._attachEventListeners();
+    }
 }
+
+export {ITEM_REGION};
