--- templates/checklist/item_row.mustache
@@ -1,67 +1,0 @@
-{{!
-    This file is part of Moodle - https://moodle.org/
-
-    Moodle is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    Moodle is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with Moodle.  If not, see <https://www.gnu.org/licenses/>.
-}}
-{{!
-    @template mod_bookit/checklist/item_row
-    Base item row for checklist table view
-
-    This is a minimal base template. Specializations should create their own
-    templates with specific styling and fields.
-
-    Context variables:
-    * id - Item ID
-    * name - Item name
-    * description - Item description (optional)
-    * categoryid - Category ID
-    * sortorder - Sort order
-
-    Example context (json):
-    {
-        "id": 1,
-        "name": "Projector",
-        "description": "Example item description",
-        "categoryid": 1,
-        "sortorder": 0
-    }
-}}
-<tr class="checklist-item-row"
-    id="checklist-item-row-{{id}}"
-    data-region="checklist-item-row"
-    data-itemid="{{id}}"
-    data-item-name="{{name}}"
-    data-item-description="{{description}}"
-    data-item-categoryid="{{categoryid}}"
-    data-item-sortorder="{{sortorder}}"
-    draggable="true">
-
-    <td class="py-2">
-        <span data-field="item-name">{{{name}}}</span>
-        {{#description}}
-        <small class="text-muted d-block" data-field="item-description">{{{description}}}</small>
-        {{/description}}
-    </td>
-
-    <td class="text-center py-2">
-        <button type="button"
-                id="edit-item-{{id}}"
-                class="btn btn-sm btn-light rounded-pill"
-                value="{{id}}"
-                data-action="edit-item"
-                data-itemid="{{id}}">
-            {{#pix}}i/edit, core{{/pix}}
-        </button>
-    </td>
-</tr>
